dircolors!
This one was real awkward. OMZ needs to be told to use GNU ls using a
zstyle command. On top of that, it also needs to see gls in the PATH,
not just ls.

This causes it to set a GNU-compatible alias with `--color=tty`.
Interestingly, the BSD flag `-G` is supported by GNU ls but has a
completely different meaning: it disables the printing of group names.